% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{accuracy}
\alias{accuracy}
\title{Compute Gaze Accuracy}
\usage{
accuracy(x, y, target_x_deg, target_y_deg, central_tendency_fun = mean)
}
\arguments{
\item{x}{Gaze azimuth in degrees.}

\item{y}{Gaze elevation in degrees.}

\item{target_x_deg}{Target azimuth in degrees.}

\item{target_y_deg}{Target elevation in degrees.}

\item{central_tendency_fun}{Function to compute central tendency (default: \code{mean}).}
}
\value{
A list with \code{offset}, \code{offset_x}, and \code{offset_y}, the total, horizontal and vertical offset of gaze from the target (in degrees).
}
\description{
Calculates the angular offset between gaze and target directions.
}
\examples{
accuracy(c(1, 2), c(1, 2), 0, 0)
}
