% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{bcea}
\alias{bcea}
\title{Bivariate Contour Ellipse Area (BCEA)}
\usage{
bcea(azi, ele, P = 0.68)
}
\arguments{
\item{azi}{Azimuth values in degrees.}

\item{ele}{Elevation values in degrees.}

\item{P}{Cumulative probability (default: 0.68).}
}
\value{
A list with the BCEA (\code{area}) and additional info about the BCEA ellipse: \code{orientation}, \code{ax1}, \code{ax2}, and \code{aspect_ratio}.
}
\description{
Computes BCEA and ellipse parameters for gaze precision.
}
\examples{
bcea(rnorm(100), rnorm(100))
}
