% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{data_loss_from_expected}
\alias{data_loss_from_expected}
\title{Compute Data Loss from Expected Sample Count}
\usage{
data_loss_from_expected(x, y, duration, frequency)
}
\arguments{
\item{x}{Azimuth values.}

\item{y}{Elevation values.}

\item{duration}{Duration in seconds.}

\item{frequency}{Sampling frequency in Hz.}
}
\value{
Percentage of data loss.
}
\description{
Calculates data loss based on expected number of samples.
}
\examples{
data_loss_from_expected(c(1, NA, 3), c(1, 2, NA), duration = 1, frequency = 3)
}
