% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{data_loss_from_invalid}
\alias{data_loss_from_invalid}
\title{Compute Data Loss from number of invalid samples.}
\usage{
data_loss_from_invalid(a, b)
}
\arguments{
\item{a}{Horizontal gaze values (e.g. azimuth or horizontal coordinate in pixels or mm).}

\item{b}{Vertical gaze values (e.g. azimuth or horizontal coordinate in pixels or mm).}
}
\value{
Percentage of missing samples.
}
\description{
Calculates percentage of missing gaze samples.
}
\examples{
data_loss_from_invalid(c(1, NA, 3), c(1, 2, NA))
}
