% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{effective_frequency}
\alias{effective_frequency}
\title{Compute Effective Sampling Frequency}
\usage{
effective_frequency(x, y, duration)
}
\arguments{
\item{x}{Azimuth values.}

\item{y}{Elevation values.}

\item{duration}{Duration in seconds.}
}
\value{
Effective frequency in Hz.
}
\description{
Calculates effective frequency based on valid samples.
}
\examples{
effective_frequency(c(1, NA, 3), c(1, 2, NA), duration = 1)
}
