% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{ETDQ_version}
\alias{ETDQ_version}
\title{Get ETDQualitizer Version}
\usage{
ETDQ_version()
}
\value{
A character string representing the version number.
}
\description{
Returns the current version string of the ETDQualitizer tool.
}
\examples{
ETDQ_version()

}
