% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{Fick_to_vector}
\alias{Fick_to_vector}
\title{Convert Fick Angles to 3D Vector}
\usage{
Fick_to_vector(azi, ele, rho = 1)
}
\arguments{
\item{azi}{Azimuth angle in degrees.}

\item{ele}{Elevation angle in degrees.}

\item{rho}{Radius (default is 1.0).}
}
\value{
A list with components \code{x}, \code{y}, and \code{z}.
}
\description{
Converts azimuth and elevation angles (in degrees) to a 3D unit vector.
}
\examples{
Fick_to_vector(30, 10)
}
