% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{rms_s2s}
\alias{rms_s2s}
\title{RMS of Sample-to-Sample Differences}
\usage{
rms_s2s(azi, ele, central_tendency_fun = mean)
}
\arguments{
\item{azi}{Azimuth values in degrees.}

\item{ele}{Elevation values in degrees.}

\item{central_tendency_fun}{Function to compute central tendency (default: \code{mean}).}
}
\value{
A list with \code{rms}, \code{rms_azi}, and \code{rms_ele}, the total RMS of sample-to-sample distances and that of the azimuthal and elevation components (all in degrees).
}
\description{
Computes root mean square of differences between successive gaze samples.
}
\examples{
rms_s2s(c(1, 2, 3), c(1, 2, 3))
}
