% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{std}
\alias{std}
\title{Standard Deviation of Gaze Samples}
\usage{
std(azi, ele)
}
\arguments{
\item{azi}{Azimuth values in degrees.}

\item{ele}{Elevation values in degrees.}
}
\value{
A list with \code{std}, \code{std_a}, and \code{std_e}, the total STD of sample-to-sample distances and that of the azimuthal and elevation components (all in degrees).
}
\description{
Computes standard deviation of azimuth and elevation.
}
\examples{
std(c(1, 2, 3), c(1, 2, 3))
}
