% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{vector_to_Fick}
\alias{vector_to_Fick}
\title{Convert 3D Vector to Fick Angles}
\usage{
vector_to_Fick(x, y, z)
}
\arguments{
\item{x}{X component of the vector.}

\item{y}{Y component of the vector.}

\item{z}{Z component of the vector.}
}
\value{
A list with components \code{azi} and \code{ele}.
}
\description{
Converts a 3D vector to azimuth and elevation angles (in degrees).
}
\examples{
vector_to_Fick(0.5, 0.2, 0.8)
}
