% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETDQualitizer.R
\name{data_loss}
\alias{data_loss}
\title{Compute Data Loss Percentage}
\usage{
data_loss(x, y)
}
\arguments{
\item{x}{Azimuth values.}

\item{y}{Elevation values.}
}
\value{
Percentage of missing samples.
}
\description{
Calculates percentage of missing gaze samples.
}
\examples{
data_loss(c(1, NA, 3), c(1, 2, NA))
}
